backend/requirements.txt
Django==4.2.0
django-debug-toolbar==4.1.0
psycopg2-binary>=2.8,<2.9
dj-database-url==0.4.1
ipython==5.3.0
subliminal==2.1.0
gunicorn==19.9.0
clint==0.5.1 
django-cors-headers==4.0.0
djangorestframework==3.14.0
dj-rest-auth[with_social]==4.0.1
django-redis==4.12.1 
pycodestyle==2.6.0
sentry-sdk==0.14.4
celery==4.4.6
redis==3.4.1
ffsubsync==0.4.3
requests==2.31.0
django-filebrowser==4.0.3
django-grappelli==3.0.3
